add np.float32

+++ rnn.example.py
@@ -24,11 +24,11 @@ x_seq = T.matrix('input')
 y_hat = T.matrix('target')
 
 
-Wi = theano.shared(np.eye(N_INPUT,N_HIDDEN), name = 'wi')
-bh = theano.shared(np.zeros(N_HIDDEN), name = 'b2_g')
-Wo = theano.shared(np.eye(N_HIDDEN,N_OUTPUT), name='Wo')
-bo = theano.shared(np.zeros(N_OUTPUT), name='bo')
-Wh = theano.shared(np.eye(N_HIDDEN,N_HIDDEN), name='Wh')
+Wi = theano.shared(np.eye(N_INPUT,N_HIDDEN).astype(np.float32), name = 'wi')
+bh = theano.shared(np.zeros(N_HIDDEN).astype(np.float32), name = 'b2_g')
+Wo = theano.shared(np.eye(N_HIDDEN,N_OUTPUT).astype(np.float32), name='Wo')
+bo = theano.shared(np.zeros(N_OUTPUT).astype(np.float32), name='bo')
+Wh = theano.shared(np.eye(N_HIDDEN,N_HIDDEN).astype(np.float32), name='Wh')
 #Wi,bh,Wo,bo,Wh = TestParser.load_matrix(fname = "rnn_parameter.txt")
 '''
 Wi = TestParser.load_matrix(N_INPUT, N_HIDDEN, name='Wi')
